examples/chat-your-pdf/requirements.txt
langchain
redis==4.5.3
# redisvl==0.0.3
python-dotenv==0.21.0
streamlit==1.24.1
streamlit-chat
google-cloud-aiplatform>=1.26.
sentence-transformers==2.2.2
altair==4.0
pypdf
